Added error types. Moved GracefulErrors to errors folder
