+++ src/types/errors/CustomErrors.ts
@@ -1,0 +1,19 @@
+export class BadRequestError extends Error {
+  name: string;
+  code: number;
+  constructor(message: string) {
+    super(message);
+    this.name = 'BadRequest';
+    this.code = 400;
+  }
+}
+
+export class UnprocessableEntityError extends Error {
+  name: string;
+  code: number;
+  constructor(message: string) {
+    super(message);
+    this.name = 'UnprocessableEntity';
+    this.code = 422;
+  }
+}
