+++ src/types/errors/GracefulError.ts
